perf: split ring into separate file
